requirements.txt
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
nltk==3.8.1
faiss-cpu==1.8.0
packaging==24.0
